# citicup-admin/app/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  model.Companies:
    items:
      properties:
        companyName:
          type: string
        id:
          type: integer
      type: object
    type: array
  model.Company:
    properties:
      companyName:
        type: string
      id:
        type: integer
    type: object
  model.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      userName:
        type: string
    type: object
  model.Users:
    items:
      properties:
        email:
          type: string
        id:
          type: integer
        password:
          type: string
        phone:
          type: string
        userName:
          type: string
      type: object
    type: array
host: localhost:8080
info:
  contact:
    name: API Support
  description: The Official Api Document For the Citicup
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Citicup Api
  version: "1.0"
paths:
  /api/company:
    get:
      consumes:
      - application/json
      description: Get All Companies.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Companies'
            type: object
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取所有公司
      tags:
      - Company
  /api/company/{id}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Companies'
            type: object
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 根据Id获取公司信息
      tags:
      - Company
  /api/users:
    get:
      consumes:
      - application/json
      description: Get All Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
            type: object
        "404":
          description: No Users
          schema:
            type: string
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取所有用户
      tags:
      - User
swagger: "2.0"